Move supporting functions to utilities.py
