+++ testing.py
@@ -15,11 +15,11 @@ class testcases(unittest.TestCase):
         self.assertEqual(response.status_code, 405)
 
     def test_003_is_email(self):
-        test = bot.bot.check_email_syntax("notanemail")
+        test = bot.utilities.check_email_syntax("notanemail")
         self.assertFalse(test)
 
     def test_004_is_email(self):
-        test = bot.bot.check_email_syntax("somename@cisco.com")
+        test = bot.utilities.check_email_syntax("somename@cisco.com")
         self.assertTrue(test)
 
     def test_005_is_cisco_email(self):
